1.0.1a2-1215.20122016
Changed the way parameters are explained
Changed the order on which online and unique users are displayed
Added a command and log counter
osu command now no longer requires a mode argument (but can be
specified)
Completely remake of the quote command
Changed how roles are done.
Fixed a bug in the yandere command where trigger wasn't getting updated
properly.
Fixed a bug where the shutdown command was not working properly.
Internal changes on the Anime command.