Alternate responses added
Alternate responses are all punched in, separated by `, for the ones with different forms (e.g., for Zygarde Complete Forme, the following responses are all accepted: Zygarde Complete, Zygarde Complete Form, Zygarde Complete Forme). Deoxys def form has unusually many responses to account for both American and British spellings (Deoxys Defence/se; Deoxys Defence/se Forme; Deoxys Defence/se Form). The only time a form without specification is accepted as an answer is when a form is considered to be the original/default form. Any form that's put up as the main Pokémon artwork in Bulbapedia is considered to be default. e.g. in this case http://bulbapedia.bulbagarden.net/wiki/Meloetta_(Pok%C3%A9mon), Meloetta Aria Forme is considered to be the original/default form, so "Meloetta" would be acceptable *only* for the Aria Form(e) and *not* for the Pirouette Form(e).