Update JDA (removes SnowflakeReference)
This will let us test ShenandoahGC again: We saw great results, but it was hindered by the STW pauses to clean up WeakReferences; with the removal of SnowflakeReference the amount of WeakReferences is reduces from dozens of millions to a few thousand per GC cycle, which will be *WAY* better for Shenandoah.

This is too long of a description for a commit with a single line change.
Yes.