Fix(roleinfo): remove "Members" as we can't track this anymore
Without a permanent cache, anything that does Guild#getMembers will just
get the amount of members we have in the cache, but there's an issue with
this: we expire members on cache.
The issue with this is that in a role with 9,000+ users, Mantaro was (wrongly)
reporting 57 members, which makes no sense.
Remove this to avoid confusion.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>