Add(profile): Make profile embed better
The old profile embed is not good on mobile, and spacing
gets hard to read even on desktop sometimes. This increases
spacing both for headers and items, using half-width spaces
(else discord will be a meme and just show one space).
Also change the badges text on profile from Top 5 Badges
to Badges Collection (Top 5) to match the Items Collection (Top 10)
text.
Footer now shows the image of whoever called the command, while
the author (header) shows the image of the person the command
was called on.
Money icon was changed to make it different to Legacy Credits,
so people can have a visual cue that they're different.
The max amount of items it'll show now it's 10, as in mobile
15 was clipping into 2 different bars, with the second bar having
a single item (which was annoying me a lot)
As with the rest of the embeds on the bot after 6.0, use the role
color as a visual cue.

As small changes, this commit also changes the appearance of the
pet status command, and the appearance of the shard command.
Those two are fairly small changes, and mostly just bug fixes
of stuff that should have been, but isn't anymore.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>