Fix(db): Fix elements not being removed due to last commit
Turns out doing "conflict": "update" on rethinkdb doesn't
actually see a change if the object has another object
within itself ((say {"hello":1, "hello-1": 2}) and you remove
hello-1, it doesn't update, which is 10 shades of fucked.

We can still update when being sure we're updating already
existing fields, and when we're sure we're not removing anything,
which should make it less meme anyway.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>