Fix(db): update values instead of overriding when possible Add(crate): change display Fix(item): don't save a seasonal object we haven't changed
This should take into account that we CAN NOT do this when a
value in a nested object is removed. It's possible on
rethinkdb itself, by using r.literal, but the java driver
does NOT implement this properly (we can't only specify
this for nested fields), so yeah.

The crate display now merges the ItemStack together instead
of showing a ton of the same item together, which is very annoying.
Though nobody has complained about it, I just did now :)

Make sure most Badge#addBadgeIfAbsent calls use ManagedObject#saveUpdating
as some of them actually might override the player object
as the player is playing the game normally, leading to item
or money loss, which is bad. This also fixes it whenever possible.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>