Fix(db): Don't process experience if there's a command processing elsewhere
This can cause a myriad of issues we really don't want, such
as lost or duplicated items.
Keep track of the user command running status on redis so we can keep an eye
on it cross-node, instead of relying on the old single-node
communication method.

Maybe we should use RabbitMQ instead? lol

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>