Fix(birthday): make so we don't spike to 40rps on executor
Back off shard handling by a random number between 0 and 3000ms,
so we don't start handling birthdays at the exact same time for all
shards.

Back off all requests (role add/remove and send message) by 400ms
and queue them internally with said backoff, so we can avoid spiking
our requests to 40rps, which in turns avoids getting banned (eventually)
for 1 hour. This is a qol update for *me*, since I don't wanna deal
with cloudflare bans, sob.

Make so we access BirthdayCacher#getCachedBirthdays with a method
and not directly on the field, as this felt very bad and was probably
midly old code.

This also removes the GUILD_BANS intent, as having it to just *log* an
event felt kinda silly. Just remove it as I don't intend on keeping
anything but basic moderation in the long term.

This also adds logging for 429'd requests on Prometheus and updates JDA
to 4.2.0_216.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>