Refactor: refactor birthday commands into its own class
Also changed some internal stuff, and clarified some startup messages.
Remove the redundant UtilsCmds#dateGMT check, as we already checked if
it was a valid timezone before, adding unnecessary checks and complexity
to all time-checking stuff.
That method was also completely removed.

Make so the time command formats time in a more conventional way using
Utils#formatDate, which also uses localized dates :)
The command also shows a clock now, because... why not.

Import net.kodehawa.mantarobot.commands.* into eval, as it was annoying
to always use the fully-qualified name.

Add the node number to the initialization message on the webhook.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>