Fix(birthday): fix wrong pool usage
A single threaded pool will definitely choke with a few long-running
tasks here. A lot of the shards will have 0ms delay or will take <400ms,
but some might take longer (one took 5k ms on logs, but probably
ratelimited-fueled, which I fixed in the last string of commits), so
make it a fixed pool of 2 and 4, so it can use that (one extra, for messages) /
those (4 total, for roles) to leverage the ones that don't take long.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>