Fix(pet): Make pet help strings better and show stamina Fix(pet): Fix pet rewards able to be 0 Fix(chop): Fix money increase not being random like on all other commands Fix(ccs): Fix json output being formatted wrongly on error messages
This fixes the issue of the ambiguous pet activity strings.
We also somehow forgot to show stamina on stats, so show it now
since we have a "low stamina" message, even though it just tells
you to feed/hydrate it, which might be self-explanatory.

This commit adds different messages to chop/mine/fish pet strings
and shows the pet name on the action.

This also fixes the issue of "0 coins, 0 items" given. Just
clamp it to 1.

Also forgot to randomize chop output, oops.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>