Fix(pet): change pet remove to pet sell, as it sells the pet
We need to keep remove as an alias since people might be used to this
already, but on help it'll show as sell, to describe better what it
actually does, which is sell the price, even if not for a fair amount.
This command is meant to be used if you want to change your pet, which
I guess it's not way too often.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>