Refactor(core): refactor MantaroListener Fix(premium): don't show key duration on activation
There's a lot of stuff that was very, very wrong here.
I can't really say it all, so just see the changes.

Also, we need to stop showing the key duration on key activation
as it really, really is misleading: a key is active for as long
as the pledge is active, not for 365 days. Only if the key was
adquired for other means (see: not by patronage) the key has a
set duration, but then it's also more expensive.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>