Port(opts): port options to context
This is the first pass, and also a necessary step to adding mention
prefix to Mantaro (which this also adds).
There's probably a ton more modernizing to do to opts though. The code
is literal hell and I'd rather try to make it better now than
make my life harder later.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>