Fix(anime): fix embed looking weird in mobile and send proper timeout message
Somehow desktop and mobile formatting is cursed.
Also we just sent "we received a SocketTimeoutException" before.
That makes no sense to an user, so send a proper timeout message.

This commit also makes it so the key is sent only if the key is invalid
after the renewal period, which is 10 days. Somehow this was 20 instead
of 10.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>