Add(mine): rework how pickaxe rewards work
This buffs gem chance with higher picks and makes it so
the stats can (eventually) be seen on iteminfo.

This also makes it so the values aren't hardcoded into
infinite chains of if/else statements on the mine command
itself, since this was the oldest command without a rewrite.

Give item categories a friendly name instead of displaying
the enum name directly.

Also track cast/repair/salvage/broken items on PlayerStats, this
is to give badges later.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>