Fix(lang): fix RTL text detection
This only really happened if you used Notepad++ to translate then copy
it into crowdin. Notepad++ inserted LTR hints on the formatting of the
text (%1$s, %2$,d, etc), which fucked  up detection as a whole.
We can basically remove them to see if we can get a clean detection.

Tested and working, seemingly.

Signed-off-by: kodehawa <david.alejandro.rubio@gmail.com>