Add(slash): centralize deferring and ephemeral replies
Having them all spread out is annoying. We should be able to just call defer on any command and call it a day, since there's no guarantee that we'll meet the timeframe discord gives us to reply.

If anything, this makes it easier to mark a command as ephemeral. We can manually mark as @NoDefer if we don't want to defer a command, which is why the defer/deferEphemeral methods are still on SlashContext and co.